+++ Intern/rayx-core/src/Beamline/Objects/MatrixSource.h
@@ -14,12 +14,12 @@ class RAYX_API MatrixSource : public LightSource {
   private:
     SourcePulseType m_sourceDistributionType;  // TODO: wo muss der name angepasst werden?
 
-    double m_linearPol_0;
-    double m_linearPol_45;
-    double m_circularPol;
+    float m_linearPol_0;
+    float m_linearPol_45;
+    float m_circularPol;
 
-    double m_verDivergence;
-    double m_sourceDepth;
+    float m_verDivergence;
+    float m_sourceDepth;
 };
 
 }  // namespace RAYX
